refactor CommandExists, use it in place of command.autoDetectCommand
